add swagger and merge conflict fixes
