+++ backend/.env.example
@@ -10,3 +10,7 @@
 DATABASE_URL="postgresql://johndoe:randompassword@localhost:5432/mydb?schema=public"
 #                          ^^^^^^   ^^^^^^^^^^^^                 ^^^^
 #                         username    password          name of your database
+
+# copypaste these to .env file
+ACCESS_TOKEN_SECRET=access
+REFRESH_TOKEN_SECRET=refresh
